a bit of reorganization here
