+++ plugins/BinaryInfo/libELF/include/libELF/elf_header.h
@@ -22,7 +22,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #ifndef ELF_HEADER_20121007_H_
 #define ELF_HEADER_20121007_H_
 
-#include "elf/elf_types.h"
+#include "elf_types.h"
 
 // The ELF file header.  This appears at the start of every ELF file.
 enum {
