+++ plugins/BinaryInfo/libELF/include/libELF/elf_move.h
@@ -22,7 +22,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #ifndef ELF_MOVE_20121007_H_
 #define ELF_MOVE_20121007_H_
 
-#include "elf/elf_types.h"
+#include "elf_types.h"
 
 // Move records.
 struct elf32_move {
