+++ plugins/BinaryInfo/libELF/include/libELF/elf_sym.h
@@ -19,21 +19,29 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#ifndef ELF_VERDAUX_20121007_H_
-#define ELF_VERDAUX_20121007_H_
+#ifndef ELF_SYM_20121007_H_
+#define ELF_SYM_20121007_H_
 
-#include "elf/elf_types.h"
+#include "elf_types.h"
 
-/* Auxialiary version information.  */
+// Symbol table entry.
 
-struct elf32_verdaux {
-	elf32_word vda_name; /* Version or dependency names */
-	elf32_word vda_next; /* Offset in bytes to next verdaux entry */
+struct elf32_sym {
+	elf32_word    st_name;  // Symbol name (string tbl index)
+	elf32_addr    st_value; // Symbol value
+	elf32_word    st_size;  // Symbol size
+	unsigned char st_info;  // Symbol type and binding
+	unsigned char st_other; // Symbol visibility
+	elf32_section st_shndx; // Section index
 };
 
-struct elf64_verdaux {
-	elf64_word vda_name; /* Version or dependency names */
-	elf64_word vda_next; /* Offset in bytes to next verdaux entry */
+struct elf64_sym {
+	elf64_word    st_name;  // Symbol name (string tbl index)
+	unsigned char st_info;  // Symbol type and binding
+	unsigned char st_other; // Symbol visibility
+	elf64_section st_shndx; // Section index
+	elf64_addr    st_value; // Symbol value
+	elf64_xword   st_size;  // Symbol size
 };
 
 #endif
