+++ plugins/BinaryInfo/libELF/include/libELF/elf_syminfo.h
@@ -22,7 +22,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #ifndef ELF_SYMINFO_20121007_H_
 #define ELF_SYMINFO_20121007_H_
 
-#include "elf/elf_types.h"
+#include "elf_types.h"
 
 /* The syminfo section if available contains additional information about
    every dynamic symbol.  */
